Grenade Launcher Shoots & Reloading stops when hit
Made the reloading stop when you get hit (needs to be play tested to make sure it's playable) because you would get stuck on the reloading state when the hit montage played.

Grenade launcher now shoots projectiles & they are replicated and apply damage properly. I format the projectile class a little for this to re-use old code.