Replicated the hitscan weapons
The shooting was not exactly replicated. When creating the random point for the shots, it was calculated on the server and on the client, but we used the one on the server. So what the clients saw was not accurate. Only the shotgun is an edge case so I need to fix that on it's own