Fixed shrinking hit box bug
I couldn't locate the error exactly but I found out it was being caused when moving the hitbox. The location and rotation moving was fine but the scale got messed up. We don't actually ever change the scale when doing this for the player so I decided to remove it.